Remove SDK API module
